print names of commands is now before parsing rc-file
The reason is: if the rc-file was changed and has syntax errors,
it may be necessary to look up the command names.
But any-dl can't print the names, if it stops with error-message
on parse-error before the commands-list will be printed.

Hence, the rc-file-parsing comes after the part, where the command
list is printed (if -v option set).