Added a rule to parse empty Parser-Definitions.
This fixes a bug that popped up
as any-dl-script error-message, many lines away from the
changes, after which that error occured.
The wrong line number created confusion, making it harder
to find the problem.
This is now fixed. Empty parsers are just empty parsers now.
Before this change all the later stuff in the rc-file was eaten
as if it would be that parser.
and at some later point the lexing/parsing failed.