Resetting Linenumber of Lexer after fileread; errmsg: with filename and (now again correct) linenumber
The "wrong line numbers" in error-messages for a syntax error of the rc-files
was not because of was what mentioned in "837c2d0fcdc7938e7a4004393897d5156f117e6a".
(There jst a syntax error was triggered for empty parsers,
 but the wrong line-numbers were not fixed with that commit.)

The wrong line-number in the errmsg was coming from adding up
line-numbers for more than one rc-file.
Using more than one rc-file was relatively new feature,
and I forgot to reset the linenum for each new file.

The linenumber will NOW be reset to 1 after each of the
rc-files is read.
So, for each new file, the linenumber starts with 1.

The syntax-error-message now also contains the filename
of the rc-file, n which the syntax error was found.

This makes fixing errors easier, when more than one rc-file is used.