For Get-Save commands in rc-file, the Save for each get will be done before the next file is downloaded.
For bulk-downloads, especially for big files, this immensely reduces
the needed amount of memory, because the files will be saved early to disk.
It allso allows to have access to the downloaded files as soon as possible,
(instead of waiting for the download of all files being completed before
they will be saved).