tagselect-Functionality implemented. Semantics will change again soon!
At the moment only the first element of a doclist is shown.
Change will be: *each* top-element of the selected items in the doclist
must be shown (dump) and put into the result.

The reason is: the selected elements are always the TOP-elements,
which matched the selector-function. And all stuff below it is in the subnodes.
These will be displayed by dump.
Dump dumps in-depth as well as *each* item in the tagselection.

It must be distincted between traversing each selected node in depth
(looking at the subnodes)
and traversing the selected doclist.
The selected stuff: always Elemtn(_,_,_).
dump_html shows all of them,
but the extractors don't do that at the moment.
This is the next needed change.