download-command now also works on Url_list's
On Url-Lists the download-command does use auto-filenaming,
which means that the filename is derived from the url.
Giving a name arbiutrariliy is not possible.
This means, that for example, an URL where the file has no
file-extension is saved as such: without file-extension.
A "mv" could be used with paste/system, but using system commands
can become a security risk.