Bugfix of download-command.
The refactored code for one-function for get- and downoad-command
retrieved the BODY of the downoaded data in both cases, which in
the case of download-command yielded in reading the body back to
RAM, after it was saved on disk!
So the intended behaviour of download was crapped.

Now it's fixed, and download-command does only writes data to disk.
No mem-usage for the body/payload needed.