Parsing html-table to Match_result works, but some problems need to be solved.
The problems are: Seperation of Different tables.
Doclist can just be concatenated, but then tables aren't distinct anymore.
Array or list of Match_results would make sense, so an internal type
should be used, that is capable of seperating these entities.
(Was planned for longer time, but not necessary so far.)
No should be considered to use!