pageinfo3-parser added
It uses htmldecode before titleextract.
That does provide the correct conversion to utf-8 (which is be seen as default
for output, that makes sense).

It seems to be a candidate for replacing "pageinfo"-parser an to throw away
"pageinfo2"-parser then.
After some testing in the real world, this possibly will be done.