function "command" refactored: larger function calls pulled out.
The commands-functions are now sperated from "command"-function.
So "command"-function is much cleaner now.

Could be made even more consequent, if ALL commands are
pulled out, not only the largeer ones. (One- and two-liners too.)

This is preparation of the switch to GADT.
But it makes sense even without switch to GADT.